🚨 fix(macos): fix clippy findings from a real local Apple Silicon run
Found by running 'cargo clippy -p eltanin-macos -p eltanin-core
-p eltanin-agent --all-targets -- -D warnings' for real on this
session's M3 Max, using CARGO_TARGET_DIR to bypass the shared dev
machine's contended ~/.cargo/shared-target lock: this is the first
real clippy pass eltanin-macos's cfg(target_os = "macos") code has
ever had (ubuntu-latest CI never even type-checks it).

- peer.rs: allow clippy::similar_names on collect_peer_context's
  effective_uid/effective_gid locals — correct, precise names, false
  positive.
- lib.rs: BSDInfo::pbi_start_tvsec/pbi_start_tvusec are already u64 on
  this platform; drop the redundant u64::from() wrapping flagged by
  clippy::useless_conversion.

+++ crates/eltanin-macos/src/lib.rs
@@ -98,9 +98,10 @@ mod imp {
         // microsecond count. See this crate's module docs on why this
         // representation is not comparable to Linux's clock-tick token
         // — harmless, since nothing compares across collectors.
-        let micros = u64::from(info.pbi_start_tvsec)
+        let micros = info
+            .pbi_start_tvsec
             .saturating_mul(1_000_000)
-            .saturating_add(u64::from(info.pbi_start_tvusec));
+            .saturating_add(info.pbi_start_tvusec);
         Evidence::Present {
             value: ProcessStartToken(micros),
             source: EvidenceSource::KernelObserved,
